+++ src/components/modules/values/ValueCard.tsx
@@ -1,4 +1,5 @@
 import { ValueDTO } from '@/backend/modules/value/ValueDTO';
+import Link from 'next/link';
 
 export const ValueCard = ({ id, value }: ValueDTO) => {
   return (
@@ -10,6 +11,8 @@ export const ValueCard = ({ id, value }: ValueDTO) => {
       rounded-md
       hover:shadow-md hover:border-blue-200
       cursor-pointer
+      text-gray-800 hover:text-blue-700
+      bg-white
     "
     >
       <span>
@@ -24,9 +27,17 @@ interface IValueCardList {
 }
 
 export const ValueCardList = ({ values }: IValueCardList) => (
-  <>
+  <div className="flex flex-col space-y-2 my-2">
     {values?.map(value => (
-      <ValueCard key={value.id} {...value} />
+      <Link
+        key={value.id}
+        href={`/implementation-references/sample-values-api/${value.id}`}
+        passHref
+      >
+        <a>
+          <ValueCard {...value} />
+        </a>
+      </Link>
     ))}
-  </>
+  </div>
 );