+++ src/pages/implementation-references/sample-values-api/create.tsx
@@ -1,0 +1,8 @@
+import { NextPage } from 'next';
+import { ValueSampleApiForm } from '@/components/modules/values/ValueSampleApiForm';
+
+const CreateSampleValuesApiPage: NextPage = () => {
+  return <ValueSampleApiForm mode="create" />;
+};
+
+export default CreateSampleValuesApiPage;