+++ src/pages/implementation-references/sample-values-api/index.tsx
@@ -1,4 +1,5 @@
 import { NextPage } from 'next';
+import Link from 'next/link';
 import { useValuesResourceSampleApiClient } from '@/hooks/sample-api/values-context.hook';
 import { useEffect } from 'react';
 import { ValueCardList } from '@/components/modules/values/ValueCard';
@@ -16,6 +17,10 @@ const IndexSampleValuesApiPage: NextPage = () => {
 
   return (
     <div className="w-full flex flex-col content-center justify-center items-center">
+      Valores Encontrados
+      <Link href="/implementation-references/sample-values-api/create">
+        Create New Value
+      </Link>
       <ValueCardList values={values} />
     </div>
   );